+++ pkg/await/apps_statefulset.go
@@ -9,6 +9,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/pulumi/pulumi-kubernetes/pkg/clients"
 	"github.com/pulumi/pulumi-kubernetes/pkg/kinds"
+	"github.com/pulumi/pulumi-kubernetes/pkg/metadata"
 	"github.com/pulumi/pulumi-kubernetes/pkg/openapi"
 	"github.com/pulumi/pulumi/pkg/diag"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -155,7 +156,8 @@ func (sia *statefulsetInitAwaiter) Await() error {
 	period := time.NewTicker(10 * time.Second)
 	defer period.Stop()
 
-	return sia.await(statefulSetWatcher, podWatcher, time.After(5*time.Minute), period.C)
+	timeout := time.Duration(metadata.TimeoutSeconds(sia.config.currentInputs, 5*60)) * time.Second
+	return sia.await(statefulSetWatcher, podWatcher, time.After(timeout), period.C)
 }
 
 func (sia *statefulsetInitAwaiter) Read() error {