+++ pkg/metadata/annotations.go
@@ -26,13 +26,15 @@ const (
 
 	AnnotationPrefix = "pulumi.com/"
 
-	AnnotationAutonamed = AnnotationPrefix + "autonamed"
-	AnnotationSkipAwait = AnnotationPrefix + "skipAwait"
+	AnnotationAutonamed      = AnnotationPrefix + "autonamed"
+	AnnotationSkipAwait      = AnnotationPrefix + "skipAwait"
+	AnnotationTimeoutSeconds = AnnotationPrefix + "timeoutSeconds"
 )
 
 // Annotations for internal Pulumi use only.
 var internalAnnotationPrefixes = []string{AnnotationAutonamed}
 
+// IsInternalAnnotation returns true if the specified annotation has the `pulumi.com/` prefix, false otherwise.
 func IsInternalAnnotation(key string) bool {
 	for _, annotationPrefix := range internalAnnotationPrefixes {
 		if strings.HasPrefix(key, annotationPrefix) {
@@ -43,6 +45,7 @@ func IsInternalAnnotation(key string) bool {
 	return false
 }
 
+// SetAnnotation sets the specified key, value annotation on the provided Unstructured object.
 func SetAnnotation(obj *unstructured.Unstructured, key, value string) {
 	annotations := obj.GetAnnotations()
 	if annotations == nil {
@@ -52,12 +55,20 @@ func SetAnnotation(obj *unstructured.Unstructured, key, value string) {
 	obj.SetAnnotations(annotations)
 }
 
+// SetAnnotationTrue sets the specified annotation key to "true" on the provided Unstructured object.
 func SetAnnotationTrue(obj *unstructured.Unstructured, key string) {
 	SetAnnotation(obj, key, AnnotationTrue)
 }
 
+// IsAnnotationTrue returns true if the specified annotation has the value "true", false otherwise.
 func IsAnnotationTrue(obj *unstructured.Unstructured, key string) bool {
 	annotations := obj.GetAnnotations()
 	value := annotations[key]
 	return value == AnnotationTrue
+}
+
+// GetAnnotationValue returns the value of the specified annotation on the provided Unstructured object.
+func GetAnnotationValue(obj *unstructured.Unstructured, key string) string {
+	annotations := obj.GetAnnotations()
+	return annotations[key]
 }